[REFACTOR]: 공통 컴포넌트 common 패키지로 이동
